# docs/api.yaml
openapi: 3.0.3
info:
  title: SaaS Travel Planner API
  description: |
    Full API reference for the SaaS Travel Planner Platform.
    This covers the SuperAdmin, Admin, and Mobile (Android App) endpoints.

    ## Authentication
    Most endpoints require a `Bearer` JWT token in the Authorization header.
    Obtain tokens from `/api/mobile/login` or via NextAuth for web dashboards.

    ## Tenant Isolation
    All data is scoped by `adminId`. Every admin manages their own users and bookings.
  version: 1.0.0
  contact:
    name: SaaS Platform Support
    url: https://github.com/voip8pbx/SaaS-Platform
servers:
  - url: https://your-domain.vercel.app
    description: Production
  - url: http://localhost:3001
    description: Local Development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, agent]
        adminId:
          type: string
...
paths:
...
  /api/mobile/register:
    post:
      tags: [Mobile Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, adminId]
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                  minLength: 8
                contactNumber:
                  type: string
                adminId:
                  type: string
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User already exists
...
  /api/mobile/site-config:
    get:
      tags: [Mobile Public]
      summary: Get site configuration for mobile app
      security: []
      parameters:
        - in: query
          name: adminId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Site configuration
        "404":
          description: Config not found
  /api/mobile/profile:
    get:
      tags: [Mobile Profile]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      tags: [Mobile Profile]
      summary: Update user profile
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                contactNumber:
                  type: string
      responses:
        "200":
          description: Updated profile
...